+++ gameEngine/static/Run.js
@@ -3,29 +3,43 @@
 //TODO import some form of break_infinity.js to use DECIMAL instead of Number
 /*
 */
+import currencies from "/currencies.js";
+import upgrade from "/upgrade.js";
+import formatOutput from "/formatOutput.js"
 //LOADING SCRIPT
 onload:(openGenTab('charaGen'));
-onload:(openSuperTab('game'));
+onload:(openSuperTab('gameTab'));
 let debug = true;
+let playerFile = false;
 //container to keep track of game progress
-saveItems = new Array();
+let saveItems = new Array();
 //UI tutor container
-itemsToDraw = new Array();
+let itemsToDraw = new Array();
 //container to neatly call/update all upgrade effects
-upgrades = new Array();
+let upgrades = new Array();
 //generator container
-generators = new Array();
+let generators = new Array();
 
 
 //functions to control tabs
+
+document.getElementById("game").addEventListener("click", openSuperTab.bind("gameTab"));
+document.getElementById("upgrades").addEventListener("click", openSuperTab.bind("upgradesTab"));
+document.getElementById("apiUI").addEventListener("click", openSuperTab.bind("apiUITab"));
+
 function openSuperTab(tab) {
   var index;
   var tabNames = document.getElementsByClassName("supers");
   for (index = 0; index < tabNames.length; index++) {
     tabNames[index].style.display = "none";
   }
-  document.getElementById(tab).style.display = "block";
+  if (this) document.getElementById(this).style.display = "block";
+  else document.getElementById(tab).style.display = "block";
 }
+
+document.getElementById("charaTab").addEventListener("click", openGenTab.bind("charaGen"));
+document.getElementById("memoryTab").addEventListener("click", openGenTab.bind("memoryGen"));
+document.getElementById("apipTab").addEventListener("click", openGenTab.bind("apipGen"));
 
 function openGenTab(tab) {
   var index;
@@ -33,82 +47,10 @@ function openGenTab(tab) {
   for (index = 0; index < tabNames.length; index++) {
     tabNames[index].style.display = "none";
   }
-  document.getElementById(tab).style.display = "block";
+  if (this) document.getElementById(this).style.display = "block";
+  else document.getElementById(tab).style.display = "block";
 }
 
-
-//TODO private
-class currencies {
-  constructor(tier, refHTML, value, growth, unlocked, prestigeTarget, prestigeButton) {
-    this.tier = tier;
-    this.refHTML = refHTML;
-    this.value = value;
-    this.growth = growth;
-    this.backgroundTotal = value;
-    this.unlocked = unlocked;
-    this.prestigeAmount = 0;
-    this.prestigeTarget = prestigeTarget;
-    this.prestigeButton = prestigeButton;
-    saveItems.push(this);
-  }
-  updateValue() {
-    this.value = this.value + this.growth;
-    if (this.backgroundTotal < this.value) {this.backgroundTotal = this.value}
-    this.backgroundTotal = this.backgroundTotal + this.growth;
-  }
-  updatePrestige() {
-    this.prestigeAmount = Math.floor(Math.sqrt(this.backgroundTotal / 1E9)) * 100;
-  }
-  prestige() {
-    if (this.prestigeAmount != 0) {
-      this.prestigeTarget.value = this.prestigeTarget.value + this.prestigeAmount;
-      this.value = 100;
-      this.growth = 0;
-      this.backgroundTotal = 100;
-      this.prestigeAmount = 0;
-      this.prestigeButton.style.visibility = "hidden";
-      return true;
-    }
-    return false;
-  }
-}
-
-//generic upgrade class
-class upgrade {
-  constructor(name, currencyBuy, cost, effectStrength, effectTarget, button, costDisplay) {
-    this.name = name;
-    this.currencyBuy = currencyBuy;
-    this.cost = cost;
-    this.effectStrength = effectStrength;
-    this.effectTarget = effectTarget;
-    this.button = button;
-    this.on = false;
-    this.show = cost / 2;
-    this.costDisplay = costDisplay;
-
-    document.getElementById(costDisplay).innerHTML = formatOutput(this.cost);
-    saveItems.push(this);
-    itemsToDraw.push(this);
-    upgrades.push(this);
-  }
-  turnOn() {
-    this.on = true;
-    this.updateMulti();
-    this.effectTarget.updateGrowth();
-    this.button.style.backgroundColor = "green";
-  }
-  updateMulti() {
-    if(this.effectStrength.amount == 0) {this.effectTarget.upgradeMulti = 1;}
-    else this.effectTarget.upgradeMulti = this.effectStrength.amount;
-    this.effectTarget.updateGrowth();
-  }
-  prestigeClean() {
-    this.on = false;
-    this.updateMulti();
-    this.effectTarget.updateGrowth();
-    this.button.style.backgroundColor = "#111";
-  }
-}
 
 
 //class for generators to keep track of cost, costgrowth, and change in chara growth
@@ -160,17 +102,25 @@ class generator {
   }
 }
 
+function addUpgradeToArrays(itemToAdd) {
+  saveItems.push(itemToAdd);
+  itemsToDraw.push(itemToAdd);
+  upgrades.push(itemToAdd);
+}
+
 //BALANCEPOINT
+//TODO refactor for pretty and for save/load
 let $memory = new currencies
   ("memory", document.getElementById("memoryTotal"), 0, 0, false, null, null);
+  saveItems.push($memory); //REFACTOR?
 let $memoryLeak = new currencies
   ("memoryLeak", document.getElementById("memoryLeakTotal"), 0, 0, true, null, null);
+  saveItems.push($memoryLeak); //REFACTOR?
 let $chara = new currencies
   ("chara", document.getElementById("charaTotal"), 100, 0, true, $memory, document.getElementById("charaPrestige"));
+  saveItems.push($chara); //REFACTOR?
 
-//TODO class definitions done, rewrite variables as generators when appropriate
 // prestige layer 1 gen
-
 //BALANCEPOINT
 let $keyboards = new generator
   ($chara, $chara, 100, 1.21, "keyboardsCost", "keyboards",
@@ -216,19 +166,24 @@ let codeCleanliness = 0;
 //upgrade layer one iteration one
 let $gen11Upgrade = new upgrade
   ("upgradeGenOne1", $chara, 4000, $autoclickers, $keyboards,
-    document.getElementById("upgradeGenOne1"), "upgradeOneCost");
+    document.getElementById("upgradeGenOne1"), document.getElementById("upgradeOneCost"));
+    addUpgradeToArrays($gen11Upgrade);
 let $gen21Upgrade = new upgrade
   ("upgradeGenTwo1", $chara, 80000, $macros, $autoclickers,
-    document.getElementById("upgradeGenTwo1"), "upgradeTwoCost");
+    document.getElementById("upgradeGenTwo1"), document.getElementById("upgradeTwoCost"));
+    addUpgradeToArrays($gen21Upgrade);
 let $gen31Upgrade = new upgrade
   ("upgradeGenThree1", $chara, 1600000, $monitors, $macros,
-    document.getElementById("upgradeGenThree1"), "upgradeThreeCost");
+    document.getElementById("upgradeGenThree1"), document.getElementById("upgradeThreeCost"));
+    addUpgradeToArrays($gen31Upgrade);
 let $gen41Upgrade = new upgrade
   ("upgradeGenFour1", $chara, 32000000, $summons, $monitors,
-    document.getElementById("upgradeGenFour1"), "upgradeFourCost");
+    document.getElementById("upgradeGenFour1"), document.getElementById("upgradeFourCost"));
+    addUpgradeToArrays($gen41Upgrade);
 let $gen51Upgrade = new upgrade
   ("upgradeGenFive1", $chara, 640000000, $keyboards, $summons,
-    document.getElementById("upgradeGenFive1"), "upgradeFiveCost");
+    document.getElementById("upgradeGenFive1"), document.getElementById("upgradeFiveCost"));
+    addUpgradeToArrays($gen51Upgrade);
 
 //button activations for upgrades
 //bundled
@@ -239,7 +194,7 @@ $gen41Upgrade.button.addEventListener("click", turnUpgradeOn.bind($gen41Upgrade)
 $gen51Upgrade.button.addEventListener("click", turnUpgradeOn.bind($gen51Upgrade));
 
 
-function turnUpgradeOn(upgradeID) {
+function turnUpgradeOn() {
   if (this.currencyBuy.value >= this.cost && this.on == false) {
     this.currencyBuy.value = this.currencyBuy.value - this.cost;
     this.turnOn();
@@ -316,7 +271,7 @@ function buyOneGenerator() {
 }
 
 function prestige() {
-  currency = this;
+  let currency = this;
   if (this.prestige()) {
     itemsToDraw.forEach(prestigeUpdate)
     function prestigeUpdate(objects) {
@@ -325,14 +280,6 @@ function prestige() {
       }
     }
   }
-}
-
-//private
-function formatOutput(output) {
-  if (output >= 10000) {
-    output = output.toExponential(2);
-  }
-  return output;
 }
 
 //private
@@ -395,6 +342,7 @@ function load() {
     .then((data) => {
       console.log(data);
       if (data != 'savefile not found') {
+        playerFile = true;
         if (debug) {
           console.log('before load: ', saveItems);
         }
@@ -451,9 +399,7 @@ function bundleSavetoSend() {
 //TODO doesnt work correctly, need to iterate over all game values and replace them
 //somehow, will figure out how soon
 /*
-convert back to usable save state
-need to accept new save string as json string from server for load
-then convert (and replace current save data) with new save data
+TODO load state function called when page loads, called with savedata values if player loads a save
 */
 function parseSaveToUse(saveData) {
   let recievedData = new Array();
@@ -464,7 +410,7 @@ function parseSaveToUse(saveData) {
     saveItems[index] = item;
     console.log(item);
   }
-  delete recievedData;
+  recievedData = null;
 }
 
 //timer private?
